Removed AsyncRetry and moved logic into Retry. Removed all classes which used AsyncRetry.
