+++ resilience4j-retry/src/main/java/io/github/resilience4j/retry/Retry.java
@@ -19,6 +19,10 @@
 package io.github.resilience4j.retry;
 
 import java.util.concurrent.Callable;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionStage;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
@@ -52,6 +56,13 @@ public interface Retry {
      * @return the retry Context
      */
     Retry.Context context();
+
+    /**
+     * Creates a async retry Context.
+     *
+     * @return the async retry Context
+     */
+    Retry.AsyncContext asyncContext();
 
     /**
      * Returns the RetryConfig of this Retry.
@@ -309,6 +320,42 @@ public interface Retry {
     }
 
     /**
+     * Decorates and executes the decorated CompletionStage.
+
+     * @param scheduler execution service to use to schedule retries
+     * @param supplier the original CompletionStage
+     * @param <T> the type of results supplied by this supplier
+     * @return the decorated CompletionStage.
+     */
+    default <T> CompletionStage<T> executeCompletionStage( ScheduledExecutorService scheduler, Supplier<CompletionStage<T>> supplier){
+        return decorateCompletionStage(this, scheduler, supplier).get();
+    }
+
+    /**
+     * Decorates CompletionStageSupplier with Retry
+     *
+     * @param retry the retry context
+     * @param scheduler execution service to use to schedule retries
+     * @param supplier completion stage supplier
+     * @param <T> type of completion stage result
+     * @return decorated supplier
+     */
+    static <T> Supplier<CompletionStage<T>> decorateCompletionStage(
+            Retry retry,
+            ScheduledExecutorService scheduler,
+            Supplier<CompletionStage<T>> supplier
+    ) {
+        return () -> {
+
+            final CompletableFuture<T> promise = new CompletableFuture<>();
+            @SuppressWarnings("unchecked") final Runnable block = new AsyncRetryBlock<>(scheduler, retry.asyncContext(), supplier, promise);
+            block.run();
+
+            return promise;
+        };
+    }
+
+    /**
      * Get the Metrics of this RateLimiter.
      *
      * @return the Metrics of this RateLimiter
@@ -380,6 +427,29 @@ public interface Retry {
         void onRuntimeError(RuntimeException runtimeException);
     }
 
+    interface AsyncContext<T> {
+
+        /**
+         *  Records a successful call.
+         */
+        void onSuccess();
+
+        /**
+         * Records an failed call.
+         * @param throwable the exception to handle
+         * @return delay in milliseconds until the next try
+         */
+        long onError(Throwable throwable);
+
+        /**
+         * check the result call.
+         *
+         * @param result the  result to validate
+         * @return delay in milliseconds until the next try if the result match the predicate
+         */
+        long onResult(T result);
+    }
+
     /**
      * An EventPublisher which subscribes to the reactive stream of RetryEvents and
      * can be used to register event consumers.
@@ -394,5 +464,65 @@ public interface Retry {
 
         EventPublisher onIgnoredError(EventConsumer<RetryOnIgnoredErrorEvent> eventConsumer);
 
+    }
+
+    class AsyncRetryBlock<T> implements Runnable {
+        private final ScheduledExecutorService scheduler;
+        private final Retry.AsyncContext<T> retryContext;
+        private final Supplier<CompletionStage<T>> supplier;
+        private final CompletableFuture<T> promise;
+
+        AsyncRetryBlock(
+                ScheduledExecutorService scheduler,
+                Retry.AsyncContext<T> retryContext,
+                Supplier<CompletionStage<T>> supplier,
+                CompletableFuture<T> promise
+        ) {
+            this.scheduler = scheduler;
+            this.retryContext = retryContext;
+            this.supplier = supplier;
+            this.promise = promise;
+        }
+
+        @Override
+        public void run() {
+            final CompletionStage<T> stage;
+
+            try {
+                stage = supplier.get();
+            } catch (Throwable t) {
+                onError(t);
+                return;
+            }
+
+            stage.whenComplete((result, t) -> {
+                if (result != null) {
+                    onResult(result);
+                } else if (t != null) {
+                    onError(t);
+                }
+            });
+        }
+
+        private void onError(Throwable t) {
+            final long delay = retryContext.onError(t);
+
+            if (delay < 1) {
+                promise.completeExceptionally(t);
+            } else {
+                scheduler.schedule(this, delay, TimeUnit.MILLISECONDS);
+            }
+        }
+
+        private void onResult(T result) {
+            final long delay = retryContext.onResult(result);
+
+            if (delay < 1) {
+                promise.complete(result);
+                retryContext.onSuccess();
+            } else {
+                scheduler.schedule(this, delay, TimeUnit.MILLISECONDS);
+            }
+        }
     }
 }
