+++ resilience4j-retry/src/main/java/io/github/resilience4j/retry/internal/RetryImpl.java
@@ -18,6 +18,7 @@
  */
 package io.github.resilience4j.retry.internal;
 
+import java.util.concurrent.CompletionException;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.concurrent.atomic.LongAdder;
@@ -159,6 +160,73 @@ public class RetryImpl<T> implements Retry {
 
     }
 
+    public final class AsyncContextImpl implements Retry.AsyncContext<T> {
+
+        private final AtomicInteger numOfAttempts = new AtomicInteger(0);
+        private final AtomicReference<Throwable> lastException = new AtomicReference<>();
+
+        @Override
+        public void onSuccess() {
+            int currentNumOfAttempts = numOfAttempts.get();
+            if (currentNumOfAttempts > 0) {
+                succeededAfterRetryCounter.increment();
+                publishRetryEvent(() -> new RetryOnSuccessEvent(name, currentNumOfAttempts, lastException.get()));
+            } else {
+                succeededWithoutRetryCounter.increment();
+            }
+        }
+
+        @Override
+        public long onError(Throwable throwable) {
+            // handle the case if the completable future throw CompletionException wrapping the original exception
+            // where original exception is the the one to retry not the CompletionException
+            // for more information about exception handling in completable future check for example :
+            //https://stackoverflow.com/questions/44409962/throwing-exception-from-completablefuture
+            if (throwable instanceof CompletionException && !exceptionPredicate.test(throwable)) {
+                if (!exceptionPredicate.test(throwable.getCause())) {
+                    failedWithoutRetryCounter.increment();
+                    publishRetryEvent(() -> new RetryOnIgnoredErrorEvent(getName(), throwable));
+                    return -1;
+                }
+                return handleOnError(throwable.getCause());
+            }
+            if (!exceptionPredicate.test(throwable)) {
+                failedWithoutRetryCounter.increment();
+                publishRetryEvent(() -> new RetryOnIgnoredErrorEvent(getName(), throwable));
+                return -1;
+            }
+            return handleOnError(throwable);
+
+        }
+
+        private long handleOnError(Throwable throwable) {
+            lastException.set(throwable);
+            int attempt = numOfAttempts.incrementAndGet();
+            if (attempt >= maxAttempts) {
+                failedAfterRetryCounter.increment();
+                publishRetryEvent(() -> new RetryOnErrorEvent(name, attempt, throwable));
+                return -1;
+            }
+
+            long interval = intervalFunction.apply(attempt);
+            publishRetryEvent(() -> new RetryOnRetryEvent(getName(), attempt, throwable, interval));
+            return interval;
+        }
+
+        @Override
+        public long onResult(T result) {
+            if (null != resultPredicate && resultPredicate.test(result)) {
+                int attempt = numOfAttempts.incrementAndGet();
+                if (attempt >= maxAttempts) {
+                    return -1;
+                }
+                return intervalFunction.apply(attempt);
+            } else {
+                return -1;
+            }
+        }
+    }
+
     /**
      * {@inheritDoc}
      */
@@ -171,6 +239,11 @@ public class RetryImpl<T> implements Retry {
     @SuppressWarnings("unchecked")
     public Context context() {
         return new ContextImpl();
+    }
+
+    @Override
+    public AsyncContext asyncContext() {
+        return new AsyncContextImpl();
     }
 
     @Override
