+++ resilience4j-retry/src/test/java/io/github/resilience4j/retry/internal/AsyncRetryTest.java
@@ -1,7 +1,11 @@
 package io.github.resilience4j.retry.internal;
 
-import static io.github.resilience4j.retry.utils.AsyncUtils.awaitResult;
-import static java.util.concurrent.CompletableFuture.completedFuture;
+import org.assertj.core.api.Assertions;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Test;
+import org.mockito.BDDMockito;
+import org.mockito.Mockito;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
@@ -13,17 +17,13 @@ import java.util.function.Supplier;
 
 import javax.xml.ws.WebServiceException;
 
-import org.assertj.core.api.Assertions;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
-import org.mockito.BDDMockito;
-import org.mockito.Mockito;
-
-import io.github.resilience4j.retry.AsyncRetry;
+import io.github.resilience4j.retry.Retry;
 import io.github.resilience4j.retry.RetryConfig;
 import io.github.resilience4j.test.AsyncHelloWorldService;
 import io.vavr.control.Try;
+
+import static io.github.resilience4j.retry.utils.AsyncUtils.awaitResult;
+import static java.util.concurrent.CompletableFuture.completedFuture;
 
 public class AsyncRetryTest {
 
@@ -41,9 +41,9 @@ public class AsyncRetryTest {
 		BDDMockito.given(helloWorldService.returnHelloWorld())
 				.willReturn(completedFuture("Hello world"));
 		// Create a Retry with default configuration
-		AsyncRetry retryContext = AsyncRetry.ofDefaults("id");
+		Retry retryContext = Retry.ofDefaults("id");
 		// Decorate the invocation of the HelloWorldService
-		Supplier<CompletionStage<String>> supplier = AsyncRetry.decorateCompletionStage(
+		Supplier<CompletionStage<String>> supplier = Retry.decorateCompletionStage(
 				retryContext,
 				scheduler,
 				() -> helloWorldService.returnHelloWorld());
@@ -64,9 +64,9 @@ public class AsyncRetryTest {
 		final RetryConfig retryConfig = RetryConfig.<String>custom().retryOnResult(s -> s.contains("NoRetry"))
 				.maxAttempts(1)
 				.build();
-		AsyncRetry retryContext = AsyncRetry.of("id", retryConfig);
+		Retry retryContext = Retry.of("id", retryConfig);
 		// Decorate the invocation of the HelloWorldService
-		Supplier<CompletionStage<String>> supplier = AsyncRetry.decorateCompletionStage(
+		Supplier<CompletionStage<String>> supplier = Retry.decorateCompletionStage(
 				retryContext,
 				scheduler,
 				() -> helloWorldService.returnHelloWorld());
@@ -98,9 +98,9 @@ public class AsyncRetryTest {
 				.willReturn(completedFuture("Hello world"));
 
 		// Create a Retry with default configuration
-		AsyncRetry retryContext = AsyncRetry.ofDefaults("id");
+		Retry retryContext = Retry.ofDefaults("id");
 		// Decorate the invocation of the HelloWorldService
-		Supplier<CompletionStage<String>> supplier = AsyncRetry.decorateCompletionStage(
+		Supplier<CompletionStage<String>> supplier = Retry.decorateCompletionStage(
 				retryContext,
 				scheduler,
 				() -> helloWorldService.returnHelloWorld());
@@ -124,9 +124,9 @@ public class AsyncRetryTest {
 				.willReturn(completedFuture("Hello world"));
 
 		// Create a Retry with default configuration
-		AsyncRetry retryContext = AsyncRetry.ofDefaults("id");
+		Retry retryContext = Retry.ofDefaults("id");
 		// Decorate the invocation of the HelloWorldService
-		Supplier<CompletionStage<String>> supplier = AsyncRetry.decorateCompletionStage(
+		Supplier<CompletionStage<String>> supplier = Retry.decorateCompletionStage(
 				retryContext,
 				scheduler,
 				() -> helloWorldService.returnHelloWorld());
@@ -160,14 +160,14 @@ public class AsyncRetryTest {
 				.willThrow(new WebServiceException("BAM!"));
 
 		// Create a Retry with default configuration
-		AsyncRetry retryContext = AsyncRetry.of(
+		Retry retryContext = Retry.of(
 				"id",
 				RetryConfig
 						.custom()
 						.maxAttempts(noOfAttempts)
 						.build());
 		// Decorate the invocation of the HelloWorldService
-		Supplier<CompletionStage<String>> supplier = AsyncRetry.decorateCompletionStage(
+		Supplier<CompletionStage<String>> supplier = Retry.decorateCompletionStage(
 				retryContext,
 				scheduler,
 				() -> helloWorldService.returnHelloWorld());
@@ -205,14 +205,14 @@ public class AsyncRetryTest {
 				.willReturn(failedFuture);
 
 		// Create a Retry with default configuration
-		AsyncRetry retryContext = AsyncRetry.of(
+		Retry retryContext = Retry.of(
 				"id",
 				RetryConfig
 						.custom()
 						.maxAttempts(noOfAttempts)
 						.build());
 		// Decorate the invocation of the HelloWorldService
-		Supplier<CompletionStage<String>> supplier = AsyncRetry.decorateCompletionStage(
+		Supplier<CompletionStage<String>> supplier = Retry.decorateCompletionStage(
 				retryContext,
 				scheduler,
 				() -> helloWorldService.returnHelloWorld());
@@ -235,7 +235,7 @@ public class AsyncRetryTest {
 
 
 		// Create a Retry with default configuration
-		AsyncRetry retryContext = AsyncRetry.of(
+		Retry retryContext = Retry.of(
 				"id",
 				RetryConfig
 						.<String>custom()
@@ -243,7 +243,7 @@ public class AsyncRetryTest {
 						.retryOnResult(s -> s.contains(retryResponse))
 						.build());
 		// Decorate the invocation of the HelloWorldService
-		Supplier<CompletionStage<String>> supplier = AsyncRetry.decorateCompletionStage(
+		Supplier<CompletionStage<String>> supplier = Retry.decorateCompletionStage(
 				retryContext,
 				scheduler,
 				() -> helloWorldService.returnHelloWorld());
