+++ resilience4j-spring-boot/src/main/java/io/github/resilience4j/retry/autoconfigure/RetryMetricsAutoConfiguration.java
@@ -15,6 +15,8 @@
  */
 package io.github.resilience4j.retry.autoconfigure;
 
+import com.codahale.metrics.MetricRegistry;
+
 import org.springframework.boot.actuate.autoconfigure.MetricRepositoryAutoConfiguration;
 import org.springframework.boot.actuate.autoconfigure.MetricsDropwizardAutoConfiguration;
 import org.springframework.boot.actuate.metrics.repository.MetricRepository;
@@ -24,11 +26,7 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import com.codahale.metrics.MetricRegistry;
-
-import io.github.resilience4j.metrics.AsyncRetryMetrics;
 import io.github.resilience4j.metrics.RetryMetrics;
-import io.github.resilience4j.retry.AsyncRetryRegistry;
 import io.github.resilience4j.retry.RetryRegistry;
 
 /**
@@ -45,13 +43,6 @@ public class RetryMetricsAutoConfiguration {
 	@Bean
 	public RetryMetrics registerRetryMetrics(RetryRegistry retryRegistry, MetricRegistry metricRegistry) {
 		RetryMetrics retryMetrics = RetryMetrics.ofRetryRegistry(retryRegistry);
-		metricRegistry.registerAll(retryMetrics);
-		return retryMetrics;
-	}
-
-	@Bean
-	public AsyncRetryMetrics asyncRetryMetrics(AsyncRetryRegistry asyncRetryRegistry, MetricRegistry metricRegistry) {
-		AsyncRetryMetrics retryMetrics = AsyncRetryMetrics.ofAsyncRetryRegistry(asyncRetryRegistry);
 		metricRegistry.registerAll(retryMetrics);
 		return retryMetrics;
 	}
