+++ resilience4j-spring-boot/src/main/java/io/github/resilience4j/retry/monitoring/endpoint/RetryEndpoint.java
@@ -19,12 +19,8 @@ package io.github.resilience4j.retry.monitoring.endpoint;
 import org.springframework.boot.actuate.endpoint.AbstractEndpoint;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Set;
 
-import io.github.resilience4j.retry.AsyncRetry;
-import io.github.resilience4j.retry.AsyncRetryRegistry;
 import io.github.resilience4j.retry.Retry;
 import io.github.resilience4j.retry.RetryRegistry;
 
@@ -36,23 +32,16 @@ import io.github.resilience4j.retry.RetryRegistry;
 public class RetryEndpoint extends AbstractEndpoint {
 
 	private final RetryRegistry retryRegistry;
-	private final AsyncRetryRegistry asyncRetryRegistry;
 
-	public RetryEndpoint(RetryRegistry retryRegistry, AsyncRetryRegistry asyncRetryRegistry) {
+	public RetryEndpoint(RetryRegistry retryRegistry) {
 		super("retries");
 		this.retryRegistry = retryRegistry;
-		this.asyncRetryRegistry = asyncRetryRegistry;
 	}
 
 	@Override
 	public RetryEndpointResponse invoke() {
-		Set<String> retriesSet = retryRegistry.getAllRetries()
-				.map(Retry::getName).sorted().toJavaSet();
-		retriesSet.addAll(asyncRetryRegistry.getAllRetries()
-				.map(AsyncRetry::getName).sorted().toJavaSet());
-
-		List<String> retries = new ArrayList<>();
-		retries.addAll(retriesSet);
+		List<String> retries = retryRegistry.getAllRetries()
+				.map(Retry::getName).sorted().toJavaList();
 		return new RetryEndpointResponse(retries);
 	}
 }
