+++ resilience4j-spring-boot/src/test/java/io/github/resilience4j/retry/RetryAutoConfigurationTest.java
@@ -15,12 +15,6 @@
  */
 package io.github.resilience4j.retry;
 
-import static io.github.resilience4j.service.test.RetryDummyService.RETRY_BACKEND_A;
-import static io.github.resilience4j.service.test.RetryDummyService.RETRY_BACKEND_B;
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.io.IOException;
-
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,6 +22,13 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
+
+import java.io.IOException;
+import java.util.Map;
+import java.util.concurrent.CompletionStage;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.TimeoutException;
 
 import io.github.resilience4j.circuitbreaker.IgnoredException;
 import io.github.resilience4j.retry.autoconfigure.RetryProperties;
@@ -37,6 +38,10 @@ import io.github.resilience4j.retry.monitoring.endpoint.RetryEventsEndpointRespo
 import io.github.resilience4j.service.test.RetryDummyService;
 import io.github.resilience4j.service.test.TestApplication;
 
+import static io.github.resilience4j.service.test.RetryDummyService.RETRY_BACKEND_A;
+import static io.github.resilience4j.service.test.RetryDummyService.RETRY_BACKEND_B;
+import static org.assertj.core.api.Assertions.assertThat;
+
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
 		classes = TestApplication.class)
@@ -44,10 +49,6 @@ public class RetryAutoConfigurationTest {
 
 	@Autowired
 	RetryRegistry retryRegistry;
-
-
-	@Autowired
-	AsyncRetryRegistry asyncRetryRegistry;
 
 	@Autowired
 	RetryProperties retryProperties;
@@ -104,4 +105,78 @@ public class RetryAutoConfigurationTest {
 		// expect aspect configured as defined in application.yml
 		assertThat(retryAspect.getOrder()).isEqualTo(399);
 	}
+
+
+	/**
+	 * The test verifies that a Async Retry instance is created and configured properly when the RetryDummyService is invoked and
+	 * that the Async Retry logic is properly handled
+	 */
+	@Test
+	public void testRetryAutoConfigurationAsync() throws Throwable {
+		assertThat(retryRegistry).isNotNull();
+
+		try {
+			final CompletionStage<String> stringCompletionStage = retryDummyService.doSomethingAsync(true);
+			String result = awaitResult(stringCompletionStage, 5);
+			assertThat(result).isNull();
+
+		} catch (IOException ex) {
+			// Do nothing. The IOException is recorded by the retry as it is one of failure exceptions
+			assertThat(ex.getMessage()).contains("Test Message");
+		}
+		// The invocation is recorded by the CircuitBreaker as a success.
+		String resultSuccess = awaitResult(retryDummyService.doSomethingAsync(false), 5);
+		assertThat(resultSuccess).isNotEmpty();
+		Retry retry = retryRegistry.retry(RETRY_BACKEND_B);
+		assertThat(retry).isNotNull();
+
+		// expect retry is configured as defined in application.yml
+		assertThat(retry.getRetryConfig().getMaxAttempts()).isEqualTo(3);
+		assertThat(retry.getName()).isEqualTo(RETRY_BACKEND_B);
+		assertThat(retry.getRetryConfig().getExceptionPredicate().test(new IOException())).isTrue();
+
+		assertThat(retry.getMetrics().getNumberOfFailedCallsWithoutRetryAttempt()).isEqualTo(0);
+		assertThat(retry.getMetrics().getNumberOfFailedCallsWithRetryAttempt()).isEqualTo(1);
+		assertThat(retry.getMetrics().getNumberOfSuccessfulCallsWithoutRetryAttempt()).isEqualTo(1);
+		assertThat(retry.getMetrics().getNumberOfSuccessfulCallsWithRetryAttempt()).isEqualTo(0);
+
+		// expect retry actuator endpoint contains both retries
+		ResponseEntity<RetryEndpointResponse> retriesList = restTemplate.getForEntity("/retries", RetryEndpointResponse.class);
+		assertThat(retriesList.getBody().getRetries()).hasSize(2).containsOnly(RETRY_BACKEND_A, RETRY_BACKEND_B);
+
+		// expect retry-event actuator endpoint recorded both events
+		ResponseEntity<RetryEventsEndpointResponse> retryBackendEventList = restTemplate.getForEntity("/retries/events/" + RETRY_BACKEND_B, RetryEventsEndpointResponse.class);
+		assertThat(retryBackendEventList.getBody().getRetryEvents()).hasSize(3);
+
+		assertThat(retry.getRetryConfig().getExceptionPredicate().test(new IOException())).isTrue();
+		assertThat(retry.getRetryConfig().getExceptionPredicate().test(new IgnoredException())).isFalse();
+
+
+		// expect aspect configured as defined in application.yml
+		assertThat(retryAspect.getOrder()).isEqualTo(399);
+	}
+
+	private <T> T awaitResult(CompletionStage<T> completionStage, long timeoutSeconds) throws Throwable {
+		try {
+			return completionStage.toCompletableFuture().get(timeoutSeconds, TimeUnit.SECONDS);
+		} catch (InterruptedException | TimeoutException e) {
+			throw new AssertionError(e);
+		} catch (ExecutionException e) {
+			throw e.getCause();
+		}
+	}
+
+	private final static class HealthResponse {
+		private Map<String, Object> details;
+
+		public Map<String, Object> getDetails() {
+			return details;
+		}
+
+		public void setDetails(Map<String, Object> details) {
+			this.details = details;
+		}
+	}
+
+
 }
