+++ resilience4j-spring-boot2/src/main/resources/META-INF/spring.factories
@@ -1,7 +1,6 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
 io.github.resilience4j.retry.autoconfigure.RetryAutoConfiguration,\
 io.github.resilience4j.retry.autoconfigure.RetryMetricsAutoConfiguration,\
-io.github.resilience4j.retry.autoconfigure.AsyncRetryMetricsAutoConfiguration,\
 io.github.resilience4j.circuitbreaker.autoconfigure.CircuitBreakerAutoConfiguration,\
 io.github.resilience4j.circuitbreaker.autoconfigure.CircuitBreakerMetricsAutoConfiguration,\
 io.github.resilience4j.ratelimiter.autoconfigure.RateLimiterAutoConfiguration,\
