+++ resilience4j-spring-boot2/src/test/java/io/github/resilience4j/service/test/RetryDummyServiceImpl.java
@@ -7,7 +7,6 @@ import java.io.IOException;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CompletionStage;
 
-import io.github.resilience4j.retry.annotation.AsyncRetry;
 import io.github.resilience4j.retry.annotation.Retry;
 
 
@@ -22,7 +21,7 @@ public class RetryDummyServiceImpl implements RetryDummyService {
 		}
 	}
 
-	@AsyncRetry(name = RetryDummyService.RETRY_BACKEND_B)
+	@Retry(name = RetryDummyService.RETRY_BACKEND_B)
 	@Override
 	public CompletionStage<String> doSomethingAsync(boolean throwException) throws IOException {
 		if (throwException) {
