+++ resilience4j-spring/src/main/java/io/github/resilience4j/retry/configure/RetryConfiguration.java
@@ -21,13 +21,10 @@ import org.springframework.context.annotation.Configuration;
 
 import io.github.resilience4j.consumer.DefaultEventConsumerRegistry;
 import io.github.resilience4j.consumer.EventConsumerRegistry;
-import io.github.resilience4j.retry.AsyncRetry;
-import io.github.resilience4j.retry.AsyncRetryRegistry;
 import io.github.resilience4j.retry.Retry;
 import io.github.resilience4j.retry.RetryConfig;
 import io.github.resilience4j.retry.RetryRegistry;
 import io.github.resilience4j.retry.event.RetryEvent;
-import io.github.resilience4j.retry.internal.InMemoryAsyncRetryRegistry;
 import io.github.resilience4j.retry.internal.InMemoryRetryRegistry;
 
 /**
@@ -54,37 +51,6 @@ public class RetryConfiguration {
 		);
 		return retryRegistry;
 	}
-
-	/**
-	 * @param retryConfigurationProperties retryConfigurationProperties retry configuration spring properties
-	 * @param retryEventConsumerRegistry   the event retry registry
-	 * @return the async retry definition registry
-	 */
-	@Bean
-	public AsyncRetryRegistry asyncRetryRegistry(RetryConfigurationProperties retryConfigurationProperties, @Qualifier("asyncRetryEventConsumerRegistry") EventConsumerRegistry<RetryEvent> retryEventConsumerRegistry) {
-		AsyncRetryRegistry retryRegistry = new InMemoryAsyncRetryRegistry();
-		retryConfigurationProperties.getBackends().forEach(
-				(name, properties) -> {
-					RetryConfig retryConfig = retryConfigurationProperties.createRetryConfig(name);
-					AsyncRetry retry = retryRegistry.retry(name, retryConfig);
-					retry.getEventPublisher().onEvent(retryEventConsumerRegistry.createEventConsumer(name, properties.getEventConsumerBufferSize()));
-				}
-		);
-		return retryRegistry;
-	}
-
-
-	/**
-	 * @param retryConfigurationProperties retry configuration spring properties
-	 * @param asyncRetryRegistry           async retry in memory registry
-	 * @return the spring retry AOP aspect
-	 */
-	@Bean
-	public AsyncRetryAspect asyncRetryAspect(RetryConfigurationProperties retryConfigurationProperties,
-	                                         AsyncRetryRegistry asyncRetryRegistry) {
-		return new AsyncRetryAspect(retryConfigurationProperties, asyncRetryRegistry);
-	}
-
 
 	/**
 	 * @param retryConfigurationProperties retry configuration spring properties
